+++ Model/essBuildInc/ButterflyModerator.h
@@ -77,7 +77,7 @@ class ButterflyModerator :
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const long int) const;
   virtual std::string getLayerString(const size_t,const long int) const;
-  virtual int getCommonSurf(const size_t) const;
+  virtual int getCommonSurf(const long int) const;
 
   /// Accessor to radius
   void setRadiusX(const double R) { outerRadius=R; }
