+++ Model/photonInc/CylContainer.h
@@ -79,7 +79,7 @@ class CylContainer :
 
   virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const long int) const;
-  virtual int getCommonSurf(const size_t) const;
+  virtual int getCommonSurf(const long int) const;
   virtual std::string getLayerString(const size_t,const long int) const;
 
   void createAll(Simulation&,const attachSystem::FixedComp&);
